test: split PDF visual scaffold trimming coverage
